pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ai-hedge-fund"
version = "0.1.0"
description = "A CrewAI-based AI hedge fund scaffold with specialist agents."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "crewai>=1.10.0,<2.0.0",
  "python-dotenv>=1.0.1,<2.0.0",
  "pypdf>=5.0.0,<6.0.0",
  "yfinance>=0.2.65,<0.3.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0.0,<9.0.0",
  "ruff>=0.11.0,<0.12.0",
]

[project.scripts]
ai-hedge-fund = "ai_hedge_fund.main:run"
ai-hedge-fund-auto = "ai_hedge_fund.automation:run"
ai-hedge-fund-safe-scan = "ai_hedge_fund.safe_scan:run"

[tool.setuptools.packages.find]
where = ["src"]

[tool.crewai]
type = "crew"

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py310"
